+++ src/supports/focus-object-swf.js
@@ -2,7 +2,7 @@
 import platform from '../util/platform';
 
 // Every Environment except IE9 considers SWF objects focusable
-const result = !platform.is.TRIDENT || !platform.is.IE9;
+const result = !platform.is.IE9;
 
 export default function() {
   return result;