+++ test/functional/style.focus-within.test.js
@@ -3,7 +3,6 @@ define(function(require) {
 
   var bdd = require('intern!bdd');
   var expect = require('intern/chai!expect');
-  var pollUntil = require('intern/dojo/node!leadfoot/helpers/pollUntil');
   var makeCustomCommand = require('../helper/leadfoot-commands');
 
   var FocusWithinCommand = makeCustomCommand({
@@ -41,7 +40,7 @@ define(function(require) {
 
         .get(require.toUrl('test/pages/style.focus-within.test.html'))
         // wait until we're really initialized
-        .then(pollUntil('return window.platform'));
+        .getPlatform();
     });
 
     bdd.it('should follow into inline SVG', function() {